PRESS4-360 | added the event logs
